Windows native support, new routing API, new Timer API [PART 2]

+++ musl/include/netinet/ip.h
@@ -12,22 +12,22 @@ struct timestamp {
 	uint8_t len;
 	uint8_t ptr;
 #if __BYTE_ORDER == __LITTLE_ENDIAN
-	unsigned int flags:4;
-	unsigned int overflow:4;
+	uint8_t flags:4;
+	uint8_t overflow:4;
 #else
-	unsigned int overflow:4;
-	unsigned int flags:4;
+	uint8_t overflow:4;
+	uint8_t flags:4;
 #endif
 	uint32_t data[9];
   };
 
 struct iphdr {
 #if __BYTE_ORDER == __LITTLE_ENDIAN
-	unsigned int ihl:4;
-	unsigned int version:4;
+	uint8_t ihl:4;
+	uint8_t version:4;
 #else
-	unsigned int version:4;
-	unsigned int ihl:4;
+	uint8_t version:4;
+	uint8_t ihl:4;
 #endif
 	uint8_t tos;
 	uint16_t tot_len;
@@ -42,11 +42,11 @@ struct iphdr {
 
 struct ip {
 #if __BYTE_ORDER == __LITTLE_ENDIAN
-	unsigned int ip_hl:4;
-	unsigned int ip_v:4;
+	uint8_t ip_hl:4;
+	uint8_t ip_v:4;
 #else
-	unsigned int ip_v:4;
-	unsigned int ip_hl:4;
+	uint8_t ip_v:4;
+	uint8_t ip_hl:4;
 #endif
 	uint8_t ip_tos;
 	uint16_t ip_len;
@@ -68,11 +68,11 @@ struct ip_timestamp {
 	uint8_t ipt_len;
 	uint8_t ipt_ptr;
 #if __BYTE_ORDER == __LITTLE_ENDIAN
-	unsigned int ipt_flg:4;
-	unsigned int ipt_oflw:4;
+	uint8_t ipt_flg:4;
+	uint8_t ipt_oflw:4;
 #else
-	unsigned int ipt_oflw:4;
-	unsigned int ipt_flg:4;
+	uint8_t ipt_oflw:4;
+	uint8_t ipt_flg:4;
 #endif
 	uint32_t data[9];
 };
